refactor: extract DayLabels from Header
